naming refactor
total calories burned -> total energy burned
active calories burned -> active energy burned

calories -> kilocalories

values in long -> to double

count parameter -> value

+++ sahha-android/src/main/java/sdk/sahha/android/domain/model/dto/Vo2MaxDto.kt
@@ -5,7 +5,7 @@ import sdk.sahha.android.domain.internal_enum.RecordingMethodsHealthConnect
 
 data class Vo2MaxDto(
     val dataType: String,
-    val value: Long,
+    val value: Double,
     val source: String,
     val startDateTime: String,
     val endDateTime: String,
